Add mildly hackish work around for bug Xml::FeedPP's item description()
Item description() will sometimes return a hash instead of text. In 
Handler.pm we replace the hash ref with ''.  In 
DBI::Stories::update_rss_full_text_field we now DO NOT mark any stories 
with a zero length RSS description as having the full text in the RSS 
even if the media source is indicated as being a full text rss media 
source.