Add todo comment to Html strip
After last weeks change we're always replacing a tag with a space instead even when the stripped tag was surrounded by white space.

We should only replace the strip tags with a space if they aren't surrounded by white space already
  e.g. 'foo<bar/>baz' should strip to 'foo baz' but foo <bar/> baz shouldn't strip to 'foo  baz' e.g. there shouldn't be an additional space.

Eventually we'll fix this but since the regular expression is so complicated that I'm going to hold off.