Use ajax to populate the media source names list for autocomplete
This is an attempt to speed up page loading times. Instead of performing 
a SQL query each time the user requests a page to get the media list, we 
abstract the query out into a separate JSON call.

Because the URL of the JSON call will have the same url across a given 
dashboard, it's results will be cached as we won't have to hit the database everytime an uncached page is requested.