Change to the Perl6 given/when switch syntax from the switch case syntax.
This was necessary because the Switch cpan module is broken in 5.14.2. We use the native Perl switch feature 
if the Perl version is 5.14 or greater. The Switch module is still needed because we support perl 5.8 but we 
will be able to remove that module entirely when we drop support for perl 5.8.